test: add acceptance tests and modiify github action to run them
The acceptance testing framework here sets up a network namespace and
a rootless network in order to isolate the testing environment. Then
it creates several interfaces with known IPs and unknown MAC
addresses. It reads the MAC address of a interface and passes it as an
environment variable to the acceptance test.

A github workflow for acceptance testing was also created.