refactor: rework arp lookup functionality
The main issue with our last approach of pinging hosts and checking
the system's ARP cache is that ARP caches have a limited size. This
will result in scans over large subnets not working as anything bigger
than the cache size (net.ipv4.neigh.default.gc_thresh3) doesn't get
saved and therefore can not be compared by the old code. To work
around this we directly bind to a socket and send out ARP requests
directly to hosts in a set of IP ranges, checking whether the
responding machine has the desired MAC. This is still fairly slow for
larger subnets due to a lower bound in the milliseconds for arp.Read's
deadline functionality.

A downside of this is that we now need the NET_RAW capability to run
the provider. This shouldn't be an issue since the point of this
provider is to run once.

Additionally instead of passing functions to represent each platform,
an interface is being used. This neatly encapsulates OS specific code.

Lastly a context based timeout has been implemented, to ensure that
lookups are retried. Useful when this provider is used alongside
provisioning VMs.