feat: overhaul and improve performance of ip data source
The IP data source now asynchronously looks up IPs. It will
simultaneously search the system's ARP cache, and search the local
network for a machine with a matching MAC. Context cancellations
interrupt the process instantly, and the network scan backoff timer
will not block cancellation.

In order to implement caching the netaddr.IP datatype was extended to
communicate whether it was from the cache or not, to avoid pinging the
returned IP twice.