chore: prepare release v0.3.0
This release overhauls much of the provider's code and logic. It now
works much faster by leveraging the system's arp cache once a valid
host is found. It respects timeouts and doesn't block. And has code
coverage of 73.5% when acceptance testing. In addition the hardcoded
interface names that ended up in the last release were removed and a
proper network interface choosing field was added.